+++ ntclient/__main__.py
@@ -25,6 +25,7 @@ from ntclient import (
 from ntclient.argparser import build_subcommands
 from ntclient.utils import CLI_CONFIG
 from ntclient.utils.exceptions import SqlException
+from ntclient.utils.sql import handle_runtime_exception
 
 
 def build_arg_parser() -> argparse.ArgumentParser:
@@ -32,7 +33,6 @@ def build_arg_parser() -> argparse.ArgumentParser:
 
     arg_parser = argparse.ArgumentParser(prog=__title__)
     arg_parser.add_argument(
-        "-v",
         "--version",
         action="version",
         version="{0} cli version {1} ".format(__title__, __version__)
@@ -40,7 +40,7 @@ def build_arg_parser() -> argparse.ArgumentParser:
     )
 
     arg_parser.add_argument(
-        "-d", "--debug", action="store_true", help="enable detailed error messages"
+        "--debug", action="store_true", help="enable detailed error messages"
     )
     arg_parser.add_argument(
         "--no-pager", action="store_true", help="disable paging (print full output)"
@@ -102,23 +102,21 @@ def main(args: list = None) -> int:  # type: ignore
         exit_code, *_results = func(_parser)
     except SqlException as sql_exception:  # pragma: no cover
         print("Issue with an sqlite database: " + repr(sql_exception))
-        if CLI_CONFIG.debug:
-            raise
+        handle_runtime_exception(args, sql_exception)
     except HTTPError as http_error:  # pragma: no cover
         err_msg = "{0}: {1}".format(http_error.code, repr(http_error))
         print("Server response error, try again: " + err_msg)
-        if CLI_CONFIG.debug:
-            raise
+        handle_runtime_exception(args, http_error)
     except URLError as url_error:  # pragma: no cover
         print("Connection error, check your internet: " + repr(url_error.reason))
-        if CLI_CONFIG.debug:
-            raise
+        handle_runtime_exception(args, url_error)
     except Exception as exception:  # pylint: disable=broad-except  # pragma: no cover
-        print("Unforeseen error, run with -d for more info: " + repr(exception))
+        print("Unforeseen error, run with --debug for more info: " + repr(exception))
         print("You can open an issue here: %s" % __url__)
         print("Or send me an email with the debug output: %s" % __email__)
-        if CLI_CONFIG.debug:
-            raise
+        print("Or, run the bug report command.")
+        print()
+        handle_runtime_exception(args, exception)
     finally:
         if CLI_CONFIG.debug:
             exc_time = time.time() - start_time