+++ ntclient/services/analyze.py
@@ -7,6 +7,7 @@ Created on Sun Nov 11 23:57:03 2018
 
 import csv
 from collections import OrderedDict
+from typing import Mapping, Sequence
 
 from tabulate import tabulate
 
@@ -18,6 +19,12 @@ from ntclient import (
     NUTR_ID_KCAL,
     NUTR_ID_PROTEIN,
 )
+from ntclient.core.nutprogbar import (
+    nutrient_progress_bars,
+    print_header,
+    print_macro_bar,
+    print_nutrient_bar,
+)
 from ntclient.persistence.sql.usda.funcs import (
     sql_analyze_foods,
     sql_food_details,
@@ -27,27 +34,26 @@ from ntclient.persistence.sql.usda.funcs import (
 from ntclient.utils import CLI_CONFIG
 
 
-################################################################################
+##############################################################################
 # Foods
-################################################################################
-def foods_analyze(food_ids: set, grams: float = 0) -> tuple:
+##############################################################################
+def foods_analyze(food_ids: set, grams: float = 100) -> tuple:
     """
     Analyze a list of food_ids against stock RDA values
-    TODO: from ntclient.utils.nutprogbar import nutprogbar
-    TODO: support -t (tabular/non-visual) output flag
+    (NOTE: only supports a single food for now... add compare foods support later)
+    TODO: support flag -t (tabular/non-visual output)
+    TODO: support flag -s (scale to 2000 kcal)
     """
 
-    ################################################################################
+    ##########################################################################
     # Get analysis
-    ################################################################################
+    ##########################################################################
     raw_analyses = sql_analyze_foods(food_ids)
     analyses = {}
     for analysis in raw_analyses:
-        food_id = analysis[0]
-        if grams:
-            anl = (analysis[1], round(analysis[2] * grams / 100, 2))
-        else:
-            anl = (analysis[1], analysis[2])
+        food_id = int(analysis[0])
+        anl = (int(analysis[1]), float(round(analysis[2] * grams / 100, 2)))
+        # Add values to list
         if food_id not in analyses:
             analyses[food_id] = [anl]
         else:
@@ -59,23 +65,26 @@ def foods_analyze(food_ids: set, grams: float = 0) -> tuple:
     nutrients = sql_nutrients_overview()
     rdas = {x[0]: x[1] for x in nutrients.values()}
 
-    ################################################################################
+    ##########################################################################
     # Food-by-food analysis (w/ servings)
-    ################################################################################
+    ##########################################################################
     servings_rows = []
     nutrients_rows = []
     for food_id, nut_val_tuples in analyses.items():
+        # Print food name
         food_name = food_des[food_id][2]
+        if len(food_name) > 45:
+            food_name = food_name[:45] + "..."
         print(
-            "\n======================================\n"
+            "\n============================================================\n"
             + "==> {0} ({1})\n".format(food_name, food_id)
-            + "======================================\n"
+            + "============================================================\n"
         )
-        print("\n=========================\nSERVINGS\n=========================\n")
 
-        ################################################################################
+        ######################################################################
         # Serving table
-        ################################################################################
+        ######################################################################
+        print_header("SERVINGS")
         headers = ["msre_id", "msre_desc", "grams"]
         serving_rows = [(x[1], x[2], x[3]) for x in serving if x[0] == food_id]
         # Print table
@@ -83,57 +92,73 @@ def foods_analyze(food_ids: set, grams: float = 0) -> tuple:
         print(servings_table)
         servings_rows.append(serving_rows)
 
+        # Show refuse (aka waste) if available
         refuse = next(
             ((x[7], x[8]) for x in food_des.values() if x[0] == food_id and x[7]), None
         )
         if refuse:
-            print("\n=========================\nREFUSE\n=========================\n")
+            print_header("REFUSE")
             print(refuse[0])
             print("    ({0}%, by mass)".format(refuse[1]))
 
-        print("\n=========================\nNUTRITION\n=========================\n")
-
-        ################################################################################
-        # Nutrient table
-        ################################################################################
-        headers = ["id", "nutrient", "rda", "amount", "units"]
+        ######################################################################
+        # Nutrient colored RDA tree-view
+        ######################################################################
+        print_header("NUTRITION")
         nutrient_rows = []
+        # TODO: skip small values (<1% RDA), report as color bar if RDA is available
         for nutrient_id, amount in nut_val_tuples:
             # Skip zero values
             if not amount:
                 continue
 
+            # Get name and unit
             nutr_desc = nutrients[nutrient_id][4] or nutrients[nutrient_id][3]
             unit = nutrients[nutrient_id][2]
 
             # Insert RDA % into row
             if rdas[nutrient_id]:
-                rda_perc = str(round(amount / rdas[nutrient_id] * 100, 1)) + "%"
+                rda_perc = float(round(amount / rdas[nutrient_id] * 100, 1))
             else:
                 rda_perc = None
             row = [nutrient_id, nutr_desc, rda_perc, round(amount, 2), unit]
 
+            # Add to list
             nutrient_rows.append(row)
 
-        ################################################################################
-        # Print table
-        ################################################################################
-        table = tabulate(nutrient_rows, headers=headers, tablefmt="presto")
-        print(table)
+        # Add to list of lists
         nutrients_rows.append(nutrient_rows)
+
+        # Print view
+        # TODO: either make this function singular, or handle plural logic here
+        _food_id = list(food_ids)[0]
+        nutrient_progress_bars(
+            {_food_id: grams},
+            [(_food_id, x[0], x[1]) for x in analyses[_food_id]],
+            nutrients,
+        )
+        # TODO: make this into the `-t` or `--tabular` branch of the function
+        # headers = ["id", "nutrient", "rda %", "amount", "units"]
+        # table = tabulate(nutrient_rows, headers=headers, tablefmt="presto")
+        # print(table)
 
     return 0, nutrients_rows, servings_rows
 
 
-################################################################################
+##############################################################################
 # Day
-################################################################################
-def day_analyze(day_csv_paths: list, rda_csv_path: str = str()) -> tuple:
-    """Analyze a day optionally with custom RDAs,
-    e.g.  nutra day ~/.nutra/rocky.csv -r ~/.nutra/dog-rdas-18lbs.csv
-    TODO: Should be a subset of foods_analyze
+##############################################################################
+def day_analyze(day_csv_paths: Sequence[str], rda_csv_path: str = str()) -> tuple:
+    """Analyze a day optionally with custom RDAs, examples:
+
+       ./nutra day tests/resources/day/human-test.csv
+
+       nutra day ~/.nutra/rocky.csv -r ~/.nutra/dog-rdas-18lbs.csv
+
+    TODO: Should be a subset of foods_analyze (encapsulate/abstract/reuse code)
     """
 
+    # Get user RDAs from CSV file, if supplied
     if rda_csv_path:
         with open(rda_csv_path, encoding="utf-8") as file_path:
             rda_csv_input = csv.DictReader(
@@ -143,6 +168,7 @@ def day_analyze(day_csv_paths: list, rda_csv_path: str = str()) -> tuple:
     else:
         rdas = []
 
+    # Get daily logs from CSV file
     logs = []
     food_ids = set()
     for day_csv_path in day_csv_paths:
@@ -155,7 +181,7 @@ def day_analyze(day_csv_paths: list, rda_csv_path: str = str()) -> tuple:
                 food_ids.add(int(entry["id"]))
         logs.append(log)
 
-    # Inject user RDAs
+    # Inject user RDAs, if supplied (otherwise fall back to defaults)
     nutrients_lists = [list(x) for x in sql_nutrients_overview().values()]
     for rda in rdas:
         nutrient_id = int(rda["id"])
@@ -166,7 +192,8 @@ def day_analyze(day_csv_paths: list, rda_csv_path: str = str()) -> tuple:
                 if CLI_CONFIG.debug:
                     substr = "{0} {1}".format(_rda, _nutrient[2]).ljust(12)
                     print("INJECT RDA: {0} -->  {1}".format(substr, _nutrient[4]))
-    nutrients = {x[0]: x for x in nutrients_lists}
+    nutrients = {int(x[0]): tuple(x) for x in nutrients_lists}
+    print(nutrients)
 
     # Analyze foods
     foods_analysis = {}
@@ -181,7 +208,7 @@ def day_analyze(day_csv_paths: list, rda_csv_path: str = str()) -> tuple:
     # Compute totals
     nutrients_totals = []
     for log in logs:
-        nutrient_totals = OrderedDict()  # dict()/{} is NOT ORDERED before 3.6/3.7
+        nutrient_totals = OrderedDict()  # NOTE: dict()/{} is NOT ORDERED before 3.6/3.7
         for entry in log:
             if entry["id"]:
                 food_id = int(entry["id"])
@@ -196,112 +223,19 @@ def day_analyze(day_csv_paths: list, rda_csv_path: str = str()) -> tuple:
                         nutrient_totals[nutr_id] += nutr_val
         nutrients_totals.append(nutrient_totals)
 
-    #######
-    # Print
+    # Print results
     buffer = BUFFER_WD - 4 if BUFFER_WD > 4 else BUFFER_WD
     for analysis in nutrients_totals:
         day_format(analysis, nutrients, buffer=buffer)
     return 0, nutrients_totals
 
 
-# TODO: why not this...? nutrients: Mapping[int, tuple]
-def day_format(analysis: dict, nutrients: dict, buffer: int = 0) -> None:
+def day_format(
+    analysis: Mapping[int, float],
+    nutrients: Mapping[int, tuple],
+    buffer: int = 0,
+) -> None:
     """Formats day analysis for printing to console"""
-
-    def print_header(header: str) -> None:
-        print(CLI_CONFIG.color_default, end="")
-        print("~~~~~~~~~~~~~~~~~~~~~~~~~~~")
-        print("--> %s" % header)
-        print("~~~~~~~~~~~~~~~~~~~~~~~~~~~")
-        print(CLI_CONFIG.style_reset_all)
-
-    def print_macro_bar(
-        _fat: float, _net_carb: float, _pro: float, _kcals_max: float, _buffer: int = 0
-    ) -> None:
-        _kcals = fat * 9 + net_carb * 4 + _pro * 4
-
-        p_fat = (_fat * 9) / _kcals
-        p_car = (_net_carb * 4) / _kcals
-        p_pro = (_pro * 4) / _kcals
-
-        # TODO: handle rounding cases, tack on to, or trim off FROM LONGEST ?
-        mult = _kcals / _kcals_max
-        n_fat = round(p_fat * _buffer * mult)
-        n_car = round(p_car * _buffer * mult)
-        n_pro = round(p_pro * _buffer * mult)
-
-        # Headers
-        f_buf = " " * (n_fat // 2) + "Fat" + " " * (n_fat - n_fat // 2 - 3)
-        c_buf = " " * (n_car // 2) + "Carbs" + " " * (n_car - n_car // 2 - 5)
-        p_buf = " " * (n_pro // 2) + "Pro" + " " * (n_pro - n_pro // 2 - 3)
-        print(
-            "  "
-            + CLI_CONFIG.color_yellow
-            + f_buf
-            + CLI_CONFIG.color_blue
-            + c_buf
-            + CLI_CONFIG.color_red
-            + p_buf
-            + CLI_CONFIG.style_reset_all
-        )
-
-        # Bars
-        print(" <", end="")
-        print(CLI_CONFIG.color_yellow + "=" * n_fat, end="")
-        print(CLI_CONFIG.color_blue + "=" * n_car, end="")
-        print(CLI_CONFIG.color_red + "=" * n_pro, end="")
-        print(CLI_CONFIG.style_reset_all + ">")
-
-        # Calorie footers
-        k_fat = str(round(fat * 9))
-        k_car = str(round(net_carb * 4))
-        k_pro = str(round(pro * 4))
-        f_buf = " " * (n_fat // 2) + k_fat + " " * (n_fat - n_fat // 2 - len(k_fat))
-        c_buf = " " * (n_car // 2) + k_car + " " * (n_car - n_car // 2 - len(k_car))
-        p_buf = " " * (n_pro // 2) + k_pro + " " * (n_pro - n_pro // 2 - len(k_pro))
-        print(
-            "  "
-            + CLI_CONFIG.color_yellow
-            + f_buf
-            + CLI_CONFIG.color_blue
-            + c_buf
-            + CLI_CONFIG.color_red
-            + p_buf
-            + CLI_CONFIG.style_reset_all
-        )
-
-    def print_nute_bar(_n_id: int, amount: float, _nutrients: dict) -> tuple:
-        nutrient = _nutrients[_n_id]
-        rda = nutrient[1]
-        tag = nutrient[3]
-        unit = nutrient[2]
-        # anti = nutrient[5]
-
-        if not rda:
-            return False, nutrient
-        attain = amount / rda
-        perc = round(100 * attain, 1)
-
-        if attain >= CLI_CONFIG.thresh_over:
-            color = CLI_CONFIG.color_over
-        elif attain <= CLI_CONFIG.thresh_crit:
-            color = CLI_CONFIG.color_crit
-        elif attain <= CLI_CONFIG.thresh_warn:
-            color = CLI_CONFIG.color_warn
-        else:
-            color = CLI_CONFIG.color_default
-
-        # Print
-        detail_amount = "{0}/{1} {2}".format(round(amount, 1), rda, unit).ljust(18)
-        detail_amount = "{0} -- {1}".format(detail_amount, tag)
-        left_index = 20
-        left_pos = round(left_index * attain) if attain < 1 else left_index
-        print(" {0}<".format(color), end="")
-        print("=" * left_pos + " " * (left_index - left_pos) + ">", end="")
-        print(" {0}%\t[{1}]".format(perc, detail_amount), end="")
-        print(CLI_CONFIG.style_reset_all)
-
-        return True, perc
 
     # Actual values
     kcals = round(analysis[NUTR_ID_KCAL])
@@ -317,7 +251,7 @@ def day_format(analysis: dict, nutrients: dict, buffer: int = 0) -> None:
     fat_rda = nutrients[NUTR_ID_FAT_TOT][1]
 
     # Print calories and macronutrient bars
-    print_header("Macronutrients")
+    print_header("Macro-nutrients")
     kcals_max = max(kcals, kcals_rda)
     rda_perc = round(kcals * 100 / kcals_rda, 1)
     print(
@@ -341,10 +275,10 @@ def day_format(analysis: dict, nutrients: dict, buffer: int = 0) -> None:
 
     # Nutrition detail report
     print_header("Nutrition detail report")
-    for n_id in analysis:
-        print_nute_bar(n_id, analysis[n_id], nutrients)
-    # TODO: below
+    for nutr_id, nutr_val in analysis.items():
+        print_nutrient_bar(nutr_id, nutr_val, nutrients)
+    # TODO: actually filter and show the number of filtered fields
     print(
-        "work in progress... "
-        "some minor fields with negligible data, they are not shown here"
+        "work in progress...",
+        "some minor fields with negligible data, they are not shown here",
     )